made alot of changes

+++ gemwallets/models.py
@@ -20,7 +20,7 @@ class GemStatus(models.TextChoices):
 class GemWallet(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     user = models.OneToOneField(User, on_delete=models.CASCADE, related_name="wallet")
-    balance = models.DecimalField(max_digits=20, decimal_places=2, default=0.00)
+    balance = models.IntegerField(default=0)
     state = models.CharField(
         max_length=20, choices=GemState.choices, default=GemState.ACTIVE
     )
